requirements.txt
rich
pyyaml
gradio
requests
